refactor(Distance): 거리 관련 비즈니스 로직 클래스로 관리 및 함수 위치 수정
