+++ src/main/java/nextstep/subway/domain/Section.java
@@ -3,7 +3,8 @@ package nextstep.subway.domain;
 import javax.persistence.*;
 import java.util.stream.Stream;
 
-import static nextstep.subway.exception.ErrorMessage.*;
+import static nextstep.subway.exception.ErrorMessage.DISTANCE_CANNOT_BE_ZERO;
+import static nextstep.subway.exception.ErrorMessage.SAME_SUBWAY_SECTION_ERROR;
 
 @Entity
 public class Section {
@@ -20,8 +21,8 @@ public class Section {
     @JoinColumn(name = "down_station_id", foreignKey = @ForeignKey(name = "fk_section_down_station"))
     private Station downStation;
 
-    // TODO: Value Object 랩핑해보기
-    private int distance;
+    @Embedded
+    private Distance distance;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "line_id", foreignKey = @ForeignKey(name = "fk_section_line"))
@@ -32,18 +33,15 @@ public class Section {
     }
 
     public Section(Station upStation, Station downStation, int distance) {
-        validateSection(upStation, downStation, distance);
+        validateSection(upStation, downStation);
         this.upStation = upStation;
         this.downStation = downStation;
-        this.distance = distance;
+        this.distance = new Distance(distance);
     }
 
-    private static void validateSection(Station upStation, Station downStation, int distance) {
+    private static void validateSection(Station upStation, Station downStation) {
         if (upStation.equals(downStation)) {
             throw new IllegalArgumentException(SAME_SUBWAY_SECTION_ERROR.getMessage());
-        }
-        if (distance <= 0) {
-            throw new IllegalArgumentException(DISTANCE_CANNOT_BE_ZERO.getMessage());
         }
     }
 
@@ -56,20 +54,26 @@ public class Section {
     }
 
     public void connectUpStationToDownStation(Section section) {
-        updateDistance(section.getDistance());
+        distance.minus(section.getDistance());
         this.upStation = section.downStation;
     }
 
     public void connectDownStationToUpStation(Section section) {
-        updateDistance(section.getDistance());
+        distance.minus(section.getDistance());
         this.downStation = section.upStation;
     }
 
-    private void updateDistance(int distance) {
-        if (this.distance <= distance) {
-            throw new IllegalArgumentException(DISTANCE_BETWEEN_STATION_OVER.getMessage());
-        }
-        this.distance -= distance;
+    public boolean hasUpStation(Station station) {
+        return upStation.equals(station);
+    }
+
+    public boolean hasDownStation(Station station) {
+        return downStation.equals(station);
+    }
+
+    public void disConnectedUpSection(Section downSection) {
+        this.downStation = downSection.downStation;
+        this.distance.plus(downSection.distance);
     }
 
     public Long getId() {
@@ -84,21 +88,7 @@ public class Section {
         return downStation;
     }
 
-    public int getDistance() {
+    public Distance getDistance() {
         return distance;
-    }
-
-    public boolean hasUpStation(Station station) {
-        return upStation.equals(station);
-    }
-
-    public boolean hasDownStation(Station station) {
-        return downStation.equals(station);
-    }
-
-    public void disConnectedUpSection(Section downSection) {
-        this.downStation = downSection.downStation;
-        // TODO: 거리 관련 처리 필요
-        this.distance = distance + downSection.distance;
     }
 }
